feat: View Page for Base Design
