+++ JewelryProductionOrder/JewelryProductionOrder.csproj
@@ -16,6 +16,7 @@
   <ItemGroup>
     <Folder Include="DB Static\" />
     <Folder Include="Migrations\" />
+    <Folder Include="Views\BaseDesign\" />
     <Folder Include="wwwroot\files\" />
   </ItemGroup>
