--- JewelryProductionOrder/Views/BaseDesign/Details.cshtml
@@ -1,36 +1,0 @@
-﻿@model JewelryDesign
-
-<form method="post" class="container">
-    <div class="caption-top row">
-        <h2>Design for Jewelry</h2>
-    </div>
-    <div class="form-group">
-        <div class="row">
-            <div class="col-10">
-                <input asp-for="@Model.Jewelry.Id" hidden />
-                <div class="col-12 py-2 form-floating">
-                    <input disabled asp-for="@Model.Name" class="form-control border-0 shadow" />
-                    <label asp-for="@Model.Name" class="ms-2"></label>
-                </div>
-                <div class="col-4">
-                    @{
-                        var isImage = Model.DesignFile.EndsWith(".jpg") || Model.DesignFile.EndsWith(".jpeg") || Model.DesignFile.EndsWith(".png") || Model.DesignFile.EndsWith(".gif");
-                        if (isImage)
-                        {
-                            <img src="@Model.DesignFile" width="100%" style="border-radius:5px; border:1px solid #bbbbbb"/>
-                        }
-                        else
-                        {
-                            <a href="@Model.DesignFile" download class="btn btn-secondary">Download Design File</a>
-                        }
-                    }
-                </div>
-        </div>
-    </div>
-</form>
-<div class="row">
-    @if (User.IsInRole(SD.Role_Customer))
-    {
-        <a asp-controller="JewelryDesign" asp-action="CustomerApprove" asp-route-id="@Model.Id" class="btn btn-primary btn-sm">Approve</a>
-    }
-</div>
