+++ JewelryProductionOrder/Views/Home/Details.cshtml
@@ -1,0 +1,19 @@
+﻿@model ShoppingCart
+
+<div class="caption-top row">
+	<h2>Design Details</h2>
+</div>
+<div class="form-group">
+	<div class="row">
+		<div class="col-10">
+			<div class="col-4">
+				<img src="@Model.BaseDesign.Image" width="100%" style="border-radius:5px; border:1px solid #bbbbbb" />
+			</div>
+			<div class="col-8 py-2 text-center">
+
+				<h3 class="text-white text-uppercase">@Model.BaseDesign.Name</h3>
+
+			</div>
+		</div>
+	</div>
+
